Move `doctype.serializeContent` to htmlparser2 adapter
